backend/requirements.txt
# Homie Backend Requirements - New Clean Architecture

# Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6

# Async Support
python-socketio==5.10.0
eventlet==0.33.3

# AI Integration
google-generativeai==0.3.2

# File Operations & Monitoring
watchdog==3.0.0
pathlib2==2.3.7

# Database
sqlite3  # Built-in with Python

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1